requirements.txt
asgiref==3.6.0
Django==4.1.6
install==1.3.5
sqlparse==0.4.3
djangorestframework
django-cors-headers
djangorestframework-simplejwt